refactor: sort Styled Components in components
